files: add custom files for automated configuration
Add custom files for automated configuration. The "server", "server_port",
and "uuid" fields on the files/etc/sing-box/config.json file must be
modified accordingly after installing the image.

Make packets from client destined to an IPv4 address that is local to the
system bypass the proxy. This includes unicast DHCP Request packets with
the IPv4 daddr not being a private IPv4 address, for which we had to put
the 'udp dport 67' rule to have it bypass the proxy. This rule ensures that
all traffic from client that is destined to the router bypasses the proxy.
This covers the case that if the LAN IP address of the router is not a
private IPv4 address, it wouldn't bypass the proxy.

This makes it so that all packets that are supposed to be routed will be
tproxied, except packets with a private IPv4 address as daddr.

Regarding matching packets for the TCP-in-UDP BPF programme, use fwmark for
egress to prevent processing traffic that is not from the TCP programme.
For client, this allows traffic to a different IP address with the same TCP
port. For server, this prevents sending packet to BPF programme if the
interface has multiple IP addresses assigned and if the TCP programme
doesn't bind to all of them.

Signed-off-by: Chester A. Unal <chester.a.unal@arinc9.com>