{.config,feeds,packages}: add bsbf feed and enable certain packages
Add the bondingshouldbefree feed and disable it as we don't provide
prebuilt packages. Put the bsbf feed to the bottom of the feed list so that
bsbf is fetched after the luci feed. This is to have the packages from
other feeds ready which the bsbf packages depend on.

Add symlinks to comgt, ell, golang, htop, mptcpd, sing-box, and enable
them. Use golang from the bsbf feed instead of the SDK. Add two symlinks to
allow both SDK and bsbf packages to use golang.

Enable comgt-ncm and kmod-usb-net-huawei-cdc-ncm to support Huawei cellular
modems with NCM.

Enable htop to monitor the usage of system resources.

Enable iptables-mod-extra, iptables-mod-tproxy, kmod-ipt-extra,
kmod-ipt-tproxy, and sing-box to provide transparent proxying.

Enable KERNEL_BPF_EVENTS to support bpf_printk() on the TCP-in-UDP BPF
programme.

Enable KERNEL_MPTCP and KERNEL_MPTCP_IPV6 to provide MPTCP IPv4 and IPv6
support.

Enable mptcpd to automatically manage the endpoints for MPTCP to use.

Enable mptcpize to check via check.mptcp.dev if MPTCP is being used.

Enable tc-full, kmod-sched-act-csum, and kmod-sched-bpf to support loading
BPF programmes.

Signed-off-by: Chester A. Unal <chester.a.unal@arinc9.com>